requirements.txt
cliff
PyYAML
